+++ packages/deepsea-tools/src/utils/isNonNullable.ts
@@ -1,3 +1,6 @@
+/** 
+ * 判断一个值是否是非空值
+ */
 export function isNonNullable<T>(value: T): value is NonNullable<T> {
     return value !== null && value !== undefined
 }