+++ packages/deepsea-tools/src/utils/sleep.ts
@@ -1,12 +1,8 @@
 /**
  * 休眠指定时间
  * @param time - 休眠的毫秒数
- * @returns Promise<1>
+ * @returns {Promise<number>}
  */
-export async function sleep(time: number): Promise<1> {
-    return new Promise<1>(resolve => {
-        setTimeout(() => {
-            resolve(1)
-        }, time)
-    })
+export async function sleep(time: number): Promise<number> {
+    return new Promise<number>(resolve => setTimeout(() => resolve(time), time))
 }