+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # soda-antd
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.29.0
+    - soda-hooks@6.9.2
+
 ## 0.6.2
 
 ### Patch Changes