+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-react-router
 
+## 1.1.10
+
+### Patch Changes
+
+- soda-hooks@6.9.2
+
 ## 1.1.9
 
 ### Patch Changes